requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
scipy>=1.11.0
